multixcan.py: refactor code into library
